requirements.txt
asgiref==3.7.2
cffi==1.15.1
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
Django==4.2.5
django-extensions==3.2.3
djangorestframework==3.14.0
fonttools==4.43.0
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pycparser==2.21
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
